corrected database type

+++ data_setup.sql
@@ -35,7 +35,7 @@ CREATE TABLE items
   id int PRIMARY KEY,
   bar_id int NOT NULL,
   type VARCHAR(15) NOT NULL,
-  price decimal(5,2) NOT NULL
+  price int NOT NULL
 );
 
 -- the days of the week
